Adding GalleryAdd component and POST on client side
