+++ src/components/App/App.jsx
@@ -1,8 +1,24 @@
 import React, { useState, useEffect } from 'react';
 import GalleryList from "../GalleryList/GalleryList.jsx";
+import GalleryAdd from "../GalleryAdd/GalleryAdd.jsx";
+import axios from 'axios';
 
 function App() {
   let [images, setImages] = useState([]);
+
+  useEffect(() => {
+    getGallery()
+  }, []);
+
+  const getGallery = () => {
+    axios.get('/api/gallery').then((response) => {
+      console.log('Data:', response.data);
+      setImages(response.data);
+    }).catch((error) => {
+      console.error(error);
+      alert('There was an error loading the gallery.');
+    });
+  };
 
     return (
       <div>
@@ -10,7 +26,8 @@ function App() {
           <h1>Photo Gallery</h1>
         </header>
         <main>
-          <GalleryList images={images} setImages={setImages} />
+          <GalleryAdd getGallery={getGallery} />
+          <GalleryList images={images} setImages={setImages} getGallery={getGallery} />
         </main>
         <footer>
           <p style={{fontSize:12}}>Please note: I did have the gallery image using an img tag before implementing MUI; see commit f565b7c.</p>
